# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sahaayak",
    "description": "Local-language voice agent for discovering government schemes, scholarships, and jobs, and checking eligibility for them.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/escalations": {
      "get": {
        "tags": [
          "escalations"
        ],
        "summary": "List Tickets",
        "operationId": "list_tickets_api_escalations_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "open",
                "claimed",
                "resolved",
                "active",
                "all"
              ],
              "type": "string",
              "default": "open",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "department",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Department"
            }
          },
          {
            "name": "overdue_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overdue Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketOut"
                  },
                  "title": "Response List Tickets Api Escalations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/escalations/{ticket_id}/claim": {
      "post": {
        "tags": [
          "escalations"
        ],
        "summary": "Claim Ticket",
        "operationId": "claim_ticket_api_escalations__ticket_id__claim_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/escalations/{ticket_id}/notes": {
      "post": {
        "tags": [
          "escalations"
        ],
        "summary": "Add Ticket Note",
        "operationId": "add_ticket_note_api_escalations__ticket_id__notes_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/escalations/{ticket_id}/route": {
      "post": {
        "tags": [
          "escalations"
        ],
        "summary": "Route Ticket",
        "operationId": "route_ticket_api_escalations__ticket_id__route_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/escalations/{ticket_id}/resolve": {
      "post": {
        "tags": [
          "escalations"
        ],
        "summary": "Resolve Ticket",
        "operationId": "resolve_ticket_api_escalations__ticket_id__resolve_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ResolveRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddNoteRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AddNoteRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveRequest": {
        "properties": {
          "resolution_code": {
            "type": "string",
            "enum": [
              "answered",
              "referred",
              "no_action",
              "duplicate",
              "unreachable"
            ],
            "title": "Resolution Code",
            "default": "answered"
          },
          "note": {
            "type": "string",
            "maxLength": 2000,
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "title": "ResolveRequest"
      },
...
      "RouteRequest": {
        "properties": {
          "department": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Department"
          },
          "routing_location": {
            "type": "string",
            "maxLength": 120,
            "title": "Routing Location",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "department"
        ],
        "title": "RouteRequest"
      },
...
      "TicketNoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "actor_id": {
            "type": "string",
            "title": "Actor Id"
          },
          "actor_role": {
            "type": "string",
            "title": "Actor Role"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "actor_id",
          "actor_role",
          "text",
          "created_at"
        ],
        "title": "TicketNoteOut"
      },
      "TicketOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "caller_context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Caller Context"
          },
          "transcript_excerpt": {
            "type": "string",
            "title": "Transcript Excerpt"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "assigned_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To"
          },
          "claimed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claimed At"
          },
          "sla_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Due At"
          },
          "sla_breached": {
            "type": "boolean",
            "title": "Sla Breached",
            "default": false
          },
          "department": {
            "type": "string",
            "title": "Department"
          },
          "routing_location": {
            "type": "string",
            "title": "Routing Location"
          },
          "routing_source": {
            "type": "string",
            "title": "Routing Source"
          },
          "routing_directory_entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Directory Entry Id"
          },
          "routing_source_url": {
            "type": "string",
            "title": "Routing Source Url",
            "default": ""
          },
          "routing_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routing Verified At"
          },
          "operator_notes": {
            "items": {
              "$ref": "#/components/schemas/TicketNoteOut"
            },
            "type": "array",
            "title": "Operator Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "resolution_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Code"
          },
          "resolution_note": {
            "type": "string",
            "title": "Resolution Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "reason",
          "transcript_excerpt",
          "status",
          "department",
          "routing_location",
          "routing_source",
          "created_at"
        ],
        "title": "TicketOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}